bug fix and document update
