+++ esql.go
@@ -1,5 +1,13 @@
 package esql
 
+import (
+	"bytes"
+	"encoding/json"
+	"errors"
+
+	"github.com/xwb1989/sqlparser"
+)
+
 // Replace ...
 // esql use replace function to apply user colName replacing policy
 type Replace func(string) string
@@ -8,18 +16,30 @@ type Replace func(string) string
 // ESql is used to hold necessary information that required in parsing
 type ESql struct {
 	whiteList    map[string]int
+	replaceList  map[string]int
 	replace      Replace
 	cadence      bool
 	pageSize     int
 	bucketNumber int
 }
 
+// init ...
+// Initialize ESql struct
+// arguments:
+//     whiteListArg: the white list that prevent any selection on columns outside it
+//                   if pass in nil, esql won't any filtering
+//     replaceListArg: any selected column name on this list will be replaced by the replace
+//                   policy specified by replaceFuncArg
+//     replaceFuncArg: the policy of column name replacement
+//     cadenceArg: boolean that indicates whether to apply special handling for cadence visibility
+//     pageSizeArg: default number of documents returned for a non-aggregation query
+//     bucketNumberArg: default number of buckets returned for an aggregation query
 func (e *ESql) init(whiteListArg []string, replaceListArg []string, replaceFuncArg Replace, cadenceArg bool, pageSizeArg int, bucketNumberArg int) {
 	for _, colName := range replaceListArg {
 		e.whiteList[colName] = 1
 	}
 	for _, colName := range whiteListArg {
-		e.whiteList[colName] = 0
+		e.replaceList[colName] = 0
 	}
 
 	e.cadence = cadenceArg
@@ -41,4 +61,54 @@ func (e *ESql) init(whiteListArg []string, replaceListArg []string, replaceFuncA
 	} else {
 		e.bucketNumber = defaultBucketNumber
 	}
+}
+
+// ConvertPretty ...
+// Transform sql to elasticsearch dsl, and prettify the output json
+// usage:
+//     dsl, err := e.ConvertPretty(sql, pageParam1, pageParam2, ...)
+// arguments:
+//     sql: the sql query needs conversion in string format
+// 	   pagination: variadic arguments that indicates es search_after for pagination
+func (e *ESql) ConvertPretty(sql string, pagination ...interface{}) (dsl string, err error) {
+	dsl, err = e.Convert(sql, pagination)
+	if err != nil {
+		return dsl, err
+	}
+
+	var prettifiedDSLBytes bytes.Buffer
+	err = json.Indent(&prettifiedDSLBytes, []byte(dsl), "", "  ")
+	if err != nil {
+		return "", err
+	}
+
+	return string(prettifiedDSLBytes.Bytes()), err
+}
+
+// Convert ...
+// Transform sql to elasticsearch dsl string
+// usage:
+//     dsl, err := e.Convert(sql, pageParam1, pageParam2, ...)
+// arguments:
+//     sql: the sql query needs conversion in string format
+//     pagination: variadic arguments that indicates es search_after for pagination
+func (e *ESql) Convert(sql string, pagination ...interface{}) (dsl string, err error) {
+	stmt, err := sqlparser.Parse(sql)
+	if err != nil {
+		return "", err
+	}
+
+	//sql valid, start to handle
+	switch stmt.(type) {
+	case *sqlparser.Select:
+		dsl, err = e.convertSelect(*(stmt.(*sqlparser.Select)), pagination)
+	default:
+		err = errors.New("esql: Queries other than select not supported")
+	}
+
+	if err != nil {
+		return "", err
+	}
+
+	return dsl, nil
 }
